# packages/backend/app/openapi.yaml
openapi: 3.0.3
info:
  title: FinMind API
  version: 0.1.0
  description: API for budgeting, bills, reminders and insights
servers:
  - url: http://localhost:8000
...
paths:
  /auth/register:
    post:
      summary: Register user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
            example:
              email: user@example.com
              password: secret123
      responses:
        '201': { description: Created }
        '400':
          description: Email already used or missing fields
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: "email already used" }
  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
            example:
              email: user@example.com
              password: secret123
      responses:
        '200':
          description: Tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: "invalid credentials" }
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Missing/invalid refresh token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: "Missing Authorization Header" }
  /auth/alerts:
    get:
      summary: List recent security alerts
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/SecurityAlert'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/alerts/{alertId}/read:
    patch:
      summary: Mark a security alert as read
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: alertId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
              example: { message: "marked as read" }
        '401':
          description: Unauthorized
        '404':
          description: Alert not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      example: { error: "invalid credentials" }
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
    SecurityAlert:
      type: object
      properties:
        id: { type: integer }
        alert_type: { type: string }
        description: { type: string }
        is_read: { type: boolean }
        created_at: { type: string, format: date-time }
...